Extract the rest of legacy enums
