+++ osu.Game/Beatmaps/Legacy/LegacyEventType.cs
@@ -1,0 +1,13 @@
+namespace osu.Game.Beatmaps.Legacy
+{
+    internal enum LegacyEventType
+    {
+        Background = 0,
+        Video = 1,
+        Break = 2,
+        Colour = 3,
+        Sprite = 4,
+        Sample = 5,
+        Animation = 6
+    }
+}
