+++ osu.Game/Beatmaps/Legacy/LegacyOrigins.cs
@@ -1,0 +1,16 @@
+namespace osu.Game.Beatmaps.Legacy
+{
+    internal enum LegacyOrigins
+    {
+        TopLeft,
+        Centre,
+        CentreLeft,
+        TopRight,
+        BottomCentre,
+        TopCentre,
+        Custom,
+        CentreRight,
+        BottomLeft,
+        BottomRight
+    }
+}
