+++ osu.Game/Beatmaps/Legacy/LegacySampleBank.cs
@@ -1,0 +1,10 @@
+namespace osu.Game.Beatmaps.Legacy
+{
+    internal enum LegacySampleBank
+    {
+        None = 0,
+        Normal = 1,
+        Soft = 2,
+        Drum = 3
+    }
+}
