+++ osu.Game/Beatmaps/Legacy/LegacyStoryLayer.cs
@@ -1,0 +1,10 @@
+namespace osu.Game.Beatmaps.Legacy
+{
+    internal enum LegacyStoryLayer
+    {
+        Background = 0,
+        Fail = 1,
+        Pass = 2,
+        Foreground = 3
+    }
+}
